libero_vla_labeler/requirements.txt
h5py>=3.8.0
numpy>=1.24.0
google-generativeai>=0.8.0
openai>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
Pillow>=10.0.0
